Feat/registration

+++ admin/includes/view_all_posts.php
@@ -18,11 +18,31 @@
                     $update_to_delete_status = mysqli_query($connection, $query);
                     confirmQuery($update_to_delete_status);
                     break;
+                case 'clone':
+                    $query = "SELECT * FROM posts WHERE post_id = '{$post_to_update_id}'";
+                    $select_post_query = mysqli_query($connection, $query);
+                    
+                    while($row = mysqli_fetch_assoc($select_post_query)) {
+                        $post_category_id = $row['post_category_id'];
+                        $post_title = $row['post_title'];
+                        $post_author = $row['post_author'];
+                        $post_image = $row['post_image'];
+                        $post_content = $row['post_content'];
+                        $post_tags = $row['post_tags'];
+                        $post_comment_count = $row['post_comment_count'];
+                        $post_status = $row['post_status'];
+                        
+                        $query = "INSERT INTO posts(post_category_id, post_title, post_author, post_date, post_image, post_content, post_tags, post_comment_count, post_status)";
+                        $query .= " VALUES({$post_category_id}, '{$post_title}', '{$post_author}',  now(), '{$post_image}', '{$post_content}', '{$post_tags}', '{$post_comment_count}', '{$post_status}' )";
+
+                        $copy_query = mysqli_query($connection, $query);
+                        confirmQuery($copy_query);
+                    }
+                    break;
                 default:
                     echo '<ERROR...>';
                     break;
             }
-
         }
     }
 
@@ -32,10 +52,11 @@
 <table class="table table-bordered table-hover">
     <div id="bulkOPtionsContainer" class="col-xs-4">
         <select class="form-control" name="bulk_options" id="">
-            <option value="draft" selected disabled>Selection Options</option>
+            <option value="draft" selected>Selection Options</option>
             <option value="published">Publish</option>
             <option value="draft">Draft</option>
             <option value="delete">Delete</option>
+            <option value="clone">Clone</option>
         </select>
     </div>
     <div class="col-xs-4">
@@ -52,8 +73,9 @@
             <th>Status</th>
             <th>Image</th>
             <th>Tags</th>
-            <th>Comments</th>
+            <th>Comments Count</th>
             <th>Date</th>
+            <th>Views Count</th>
             <th>View Post</th>
             <th>Edit Post</th>
             <th>Delete Post</th>
@@ -61,7 +83,7 @@
     </thead>
     <tbody>
         <?php 
-            $query = "SELECT * FROM posts";
+            $query = "SELECT * FROM posts ORDER BY post_id DESC";
             $select_posts = mysqli_query($connection, $query);
 
             while($row = mysqli_fetch_assoc($select_posts)) {
@@ -74,6 +96,7 @@
                 $post_tags = $row['post_tags'];
                 $post_comment_count = $row['post_comment_count'];
                 $post_date = $row['post_date'];
+                $post_views_count = $row['post_views_count'];
                 
                 echo "<tr>";
                 ?>
@@ -96,9 +119,10 @@
                 echo "<td>$post_tags</td>";
                 echo "<td>$post_comment_count</td>";
                 echo "<td>$post_date</td>";
+                echo "<td><a href='posts.php?reset={$post_id}'>$post_views_count</a></td>";
                 echo "<td><a href='../post.php?p_id={$post_id}'>View</a></td>";
                 echo "<td><a href='posts.php?source=edit_post&p_id={$post_id}'>Edit</a></td>";
-                echo "<td><a href='posts.php?delete={$post_id}'>Delete</a></td>";
+                echo "<td><a href='posts.php?delete={$post_id}' onclick='return confirm('Are you sure you want to delete');'>Delete</a></td>";
                 echo "</td>";
             }
         ?>
@@ -108,6 +132,16 @@
                 $query = "DELETE FROM posts WHERE post_id = {$post_id_to_delete} ";
                 $delete_query = mysqli_query($connection, $query);
                 confirmQuery($delete_query);
+                header("Location: posts.php");
+            }
+            
+            if(isset($_GET['reset'])) {
+                $post_id_to_reset = $_GET['reset'];
+                $query = "UPDATE posts SET post_views_count = 0 WHERE post_id = ";
+                $query .= mysqli_escape_string($connection, $post_id_to_reset);
+                $reset_query = mysqli_query($connection, $query);
+                confirmQuery($reset_query);
+                header("Location: posts.php");
             }
         ?>
     </tbody>
